% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{col_12}
\alias{col_12}
\alias{col_11}
\alias{col_10}
\alias{col_9}
\alias{col_8}
\alias{col_7}
\alias{col_6}
\alias{col_5}
\alias{col_4}
\alias{col_3}
\alias{col_2}
\alias{col_1}
\title{Shorthand for column of width 12}
\usage{
col_12(...)

col_11(...)

col_10(...)

col_9(...)

col_8(...)

col_7(...)

col_6(...)

col_5(...)

col_4(...)

col_3(...)

col_2(...)

col_1(...)
}
\arguments{
\item{...}{Arguments passed to `column()`}
}
\description{
Shorthand for column of width 12

Shorthand for column of width 11

Shorthand for column of width 10

Shorthand for column of width 9

Shorthand for column of width 8

Shorthand for column of width 7

Shorthand for column of width 6

Shorthand for column of width 5

Shorthand for column of width 4

Shorthand for column of width 3

Shorthand for column of width 2

Shorthand for column of width 1
}
