% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_countries_ui.R
\name{mod_countries_ui}
\alias{mod_countries_ui}
\title{Country profile UI-side function}
\usage{
mod_countries_ui(id)
}
\arguments{
\item{id}{Internal parameter for Shiny.}
}
\description{
A shiny Module.
}
